feat: Add Gradle wrapper files for version 9.2.0
- Added gradle-wrapper.properties to configure Gradle distribution.
- Introduced gradlew shell script for Unix-based systems to execute Gradle tasks.
- Included gradlew.bat for Windows to facilitate Gradle execution.

+++ src/main/java/com/noctivag/chaos/events/ExplosionChaosEvent.java
@@ -25,7 +25,7 @@ public class ExplosionChaosEvent implements ChaosEvent {
 		double offsetY = random.nextDouble() * 3;
 		double offsetZ = (random.nextDouble() - 0.5) * 10;
 
-		BlockPos pos = targetPlayer.getBlockPos().add(offsetX, offsetY, offsetZ);
+		BlockPos pos = targetPlayer.getBlockPos().add((int)offsetX, (int)offsetY, (int)offsetZ);
 
 		float power = 1.0f + random.nextFloat() * ChaosMod.getConfig().explosionRadius;
